vulns: extract CachedFeed base class from CPANSec
Shared load/refresh/from_file for JSON feeds cached under
HOMEBREW_CACHE/vulns/: mtime-gated freshness check, download to a
per-process sibling Tempfile, validate, atomic rename, and stale-cache
fallback with a warning on failure. Subclasses implement .data_url,
.cache_filename and #initialize(data) and may override
.default_max_age. CPANSec now inherits it; behaviour unchanged apart
from the diagnostic wording naming the cache filename.
