+++ Library/Homebrew/vulns/cpan_sec.rb
@@ -1,9 +1,7 @@
 # typed: strict
 # frozen_string_literal: true
 
-require "json"
-require "tempfile"
-require "utils/curl"
+require "vulns/cached_feed"
 
 module Homebrew
   module Vulns
@@ -14,16 +12,15 @@ module Homebrew
     # keyed on CPAN distribution name. This class fetches and caches that file
     # and exposes advisories per distribution. Evaluating `affected_versions`
     # range strings against a formula version is left to {Vulns::Match}.
-    class CPANSec
-      extend Utils::Output::Mixin
-
+    class CPANSec < CachedFeed
       DATA_URL = "https://raw.githubusercontent.com/briandfoy/cpan-security-advisory/" \
                  "master/cpan-security-advisory.json"
-      CACHE_FILENAME = "cpansa.json"
-      DEFAULT_MAX_AGE = 86_400
-      private_constant :CACHE_FILENAME, :DEFAULT_MAX_AGE
 
-      class Error < RuntimeError; end
+      sig { override.returns(String) }
+      def self.data_url = DATA_URL
+
+      sig { override.returns(String) }
+      def self.cache_filename = "cpansa.json"
 
       Advisory = Struct.new(
         :id, :cves, :affected_versions, :fixed_versions,
@@ -31,57 +28,14 @@ module Homebrew
         keyword_init: true
       )
 
-      sig { params(cache: Pathname, max_age: Integer).returns(T.attached_class) }
-      def self.load(cache: HOMEBREW_CACHE/"vulns", max_age: DEFAULT_MAX_AGE)
-        cache_file = cache/CACHE_FILENAME
-        return from_file(cache_file) if cache_file.exist? && (Time.now - cache_file.mtime) <= max_age
-
-        refresh(cache_file)
-      rescue ErrorDuringExecution, Error => e
-        raise unless cache_file.exist?
-
-        opoo "Failed to refresh CPANSA data (#{e.message.lines.first&.strip}); " \
-             "using cached copy from #{cache_file.mtime}."
-        from_file(cache_file)
-      end
-
-      # Download to a per-process sibling temp file and validate before
-      # atomically replacing the cache so a failed, truncated or concurrent
-      # fetch cannot corrupt the stale copy.
-      sig { params(cache_file: Pathname).returns(T.attached_class) }
-      def self.refresh(cache_file)
-        cache_file.dirname.mkpath
-        Tempfile.create([CACHE_FILENAME, ".download"], cache_file.dirname.to_s) do |tmp|
-          tmp.close
-          path = Pathname(tmp.path)
-          Utils::Curl.curl_download("--fail", "--silent", DATA_URL, to: path)
-          loaded = from_file(path)
-          File.rename(path, cache_file)
-          return loaded
-        end
-      end
-
-      sig { params(path: Pathname).returns(T.attached_class) }
-      def self.from_file(path)
-        new(JSON.parse(path.read))
-      rescue JSON::ParserError => e
-        raise Error, "Failed to parse CPANSA data at #{path}: #{e.message}"
-      end
-
-      sig { params(data: T.anything).void }
+      sig { override.params(data: T.anything).void }
       def initialize(data)
+        super
         raise Error, "CPANSA data is not a JSON object" unless (top = as_hash(data))
         raise Error, "CPANSA data missing 'dists' key" unless (dists = as_hash(top["dists"]))
 
         @dists = T.let(dists, T::Hash[String, T.untyped])
         @meta = T.let(as_hash(top["meta"]) || {}, T::Hash[String, T.untyped])
-      end
-
-      sig { params(value: T.anything).returns(T.nilable(T::Hash[String, T.untyped])) }
-      def as_hash(value)
-        case value
-        when Hash then value
-        end
       end
 
       sig { returns(T::Hash[String, T.untyped]) }
